+++ msal/src/test/java/com/microsoft/identity/client/e2e/tests/network/AcquireTokenNetworkTest.java
@@ -24,10 +24,21 @@ package com.microsoft.identity.client.e2e.tests.network;
 
 import com.microsoft.identity.client.AcquireTokenParameters;
 import com.microsoft.identity.client.AcquireTokenSilentParameters;
+import com.microsoft.identity.client.e2e.shadows.ShadowAuthority;
+import com.microsoft.identity.client.e2e.shadows.ShadowMsalUtils;
+import com.microsoft.identity.client.e2e.shadows.ShadowStorageHelper;
+import com.microsoft.identity.client.e2e.tests.AcquireTokenAbstractTest;
+import com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper;
 import com.microsoft.identity.client.e2e.utils.ErrorCodes;
 import com.microsoft.identity.internal.testutils.TestUtils;
+import com.microsoft.identity.internal.testutils.labutils.LabUserHelper;
+import com.microsoft.identity.internal.testutils.labutils.LabUserQuery;
 
+import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestRunner;
+import org.robolectric.annotation.Config;
 
 import java.util.Arrays;
 
@@ -37,7 +48,19 @@ import static com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper.suc
 import static com.microsoft.identity.client.e2e.utils.AcquireTokenTestHelper.successfulSilentCallback;
 import static com.microsoft.identity.client.e2e.utils.RoboTestUtils.flushScheduler;
 
-public abstract class AcquireTokenNetworkTest extends AcquireTokenNetworkAbstractTest {
+@RunWith(RobolectricTestRunner.class)
+@Config(shadows = {ShadowStorageHelper.class, ShadowAuthority.class, ShadowMsalUtils.class})
+public abstract class AcquireTokenNetworkTest extends AcquireTokenAbstractTest implements IAcquireTokenNetworkTest {
+
+    String mUsername;
+
+    @Before
+    public void setup() {
+        AcquireTokenTestHelper.setAccount(null);
+        final LabUserQuery query = getLabUserQuery();
+        mUsername = LabUserHelper.loadUserForTest(query);
+        super.setup();
+    }
 
     @Test
     public void testAcquireTokenSuccess() {
